+++ packages/fleury_widgets/lib/src/range_slider.dart
@@ -127,7 +127,7 @@ class _RangeSliderState extends State<RangeSlider> {
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
-    Focus.maybeOf(context); // rebuild on focus change
+    FocusManager.maybeOf(context); // rebuild on focus change
     final registration = FormControlScope.maybeOf(context);
     if (!identical(registration, _formRegistration)) {
       _formRegistration?.release(this);